Move `FusionUtils` into `_private` package.
Remove some unused Ion helpers.
