+++ runtime/src/main/java/dev/ionfusion/fusion/FusionIterator.java
@@ -9,7 +9,7 @@ import static dev.ionfusion.fusion.FusionList.isList;
 import static dev.ionfusion.fusion.FusionList.unsafeListIterator;
 import static dev.ionfusion.fusion.FusionSexp.isSexp;
 import static dev.ionfusion.fusion.FusionSexp.unsafeSexpIterator;
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
 import com.amazon.ion.IonValue;
 import java.io.IOException;
 import java.util.Iterator;
