+++ runtime/src/main/java/dev/ionfusion/fusion/FusionValue.java
@@ -3,8 +3,8 @@
 
 package dev.ionfusion.fusion;
 
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_OBJECT_ARRAY;
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_STRING_ARRAY;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_STRING_ARRAY;
 
 import com.amazon.ion.IonValue;
 import com.amazon.ion.ValueFactory;
