+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxContainer.java
@@ -3,7 +3,7 @@
 
 package dev.ionfusion.fusion;
 
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_OBJECT_ARRAY;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
 
 abstract class SyntaxContainer
     extends SyntaxValue
