+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxGetProc.java
@@ -6,8 +6,8 @@ package dev.ionfusion.fusion;
 import static dev.ionfusion.fusion.FusionIo.safeWrite;
 import static dev.ionfusion.fusion.FusionNumber.checkIntArgToJavaInt;
 import static dev.ionfusion.fusion.FusionText.checkRequiredTextArg;
-import static dev.ionfusion.fusion.FusionUtils.writeFriendlyIndex;
 import static dev.ionfusion.fusion.FusionVoid.voidValue;
+import static dev.ionfusion.fusion._private.FusionUtils.writeFriendlyIndex;
 
 final class SyntaxGetProc
     extends Procedure
