+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxKeyword.java
@@ -3,10 +3,11 @@
 
 package dev.ionfusion.fusion;
 
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_OBJECT_ARRAY;
-import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_OBJECT_ARRAY;
+
 import com.amazon.ion.IonException;
 import com.amazon.ion.IonWriter;
+import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import java.io.IOException;
 
 final class SyntaxKeyword
