+++ runtime/src/main/java/dev/ionfusion/fusion/SyntaxList.java
@@ -7,11 +7,12 @@ import static dev.ionfusion.fusion.FusionList.immutableList;
 import static dev.ionfusion.fusion.FusionList.isImmutableList;
 import static dev.ionfusion.fusion.FusionList.nullList;
 import static dev.ionfusion.fusion.FusionList.unsafeListElement;
-import static dev.ionfusion.fusion.FusionUtils.EMPTY_STRING_ARRAY;
+import static dev.ionfusion.fusion._private.FusionUtils.EMPTY_STRING_ARRAY;
 import static java.lang.System.arraycopy;
+
+import com.amazon.ion.IonWriter;
 import dev.ionfusion.fusion.FusionList.BaseList;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
-import com.amazon.ion.IonWriter;
 import java.io.IOException;
 
 final class SyntaxList
